Experiment, drawing out calc guide
So I experimented with g.js repeatedly drawing rotated lines
and trying to get a line back to the starting point.

In i.js I figured out how to draw an arc backwards from a rotated center
point to the origin center point.

In j.js I translated my drawings in SketchPad into drawings in the
canvas showing all the intermediate lines and angles I used to calculate
dx and dy from a center point to a rotated center point. The right-angle
sign is messed up because it would be a lot of work to figure out where
the right angle sign intersected with the "P1"-"P2" line (i.e. the line
from the original center to the new center. It's late, I need to get up
early for work.

I also added some input fields to a.html and widened them so I could
add more debug info.