Slowly approaching a smooth wave
Each consecutive dot can have more of it's angle applied so that the
first dot's angle doesn't shake the whole line violently. If we apply a
tiny bit of the first dot, a little bit more of the 2nd, etc. then each
consecutive dots angle will come out more and more as it moves down the
line, _hopefully_ producing a wave.