Success: the tree is drawn properly.
I'm not great with angles so this is hard to think about, let alone
explain.

Child branches are drawn such that the parent angle becomes zero
radians. A child branch drawn at zero degrees relative to the parent
will go back in the same direction as the parent. A child branch drawn
at Pi radians will continue in the same direction as the parent. In
order to calculate where zero raians is relative to the parent I'm
taking the parent angle and rotating by Pi radians. Child branches are
spread over 2Pi radians and are spread out evenly from Pi radians. This
means child branches bunch up facing outward.

Zero radians points straight down.

If we start with a parent branch at 0 radians then the first three child
branches will be at (0.5 * pi), pi, (1.5 * pi), or left, up and right.
If we then draw three children starting at the 0.5 pi branch, then we
need branches at 0, 0.5 pi and pi, or down, left and up. This illustrates
that we subtract Pi from the parent branch and then add 0.5 * pi:
0.5 pi - pi + 0.5 pi = 0 radians: straight down again.

Above are branches with 0.5 pi spacing, which conveniently "uses up" the
full 2 * pi radians. If we used 0.25 pi radians instead then we can't
just start from 0 or we'd have 0.25 pi, 0.5 pi and 0.75 pi, or
down-left, left and up-left. In order to get the branches to point
outward we need to find out how much angle is "left over" and space that
around the branches: in this case the left-over is 1.25 pi where 0.625
comes before the first branch and after the last. This gives us the
first branch at up-left-up, up, up-right-up.