Draw sine & cosine waves and link to circle
Draws a verticlal cosine wave (undulating left and right) and a
horizontal sine wave (undulating up and down) according to the angle of
the center line. Also draw lines linking the end of the center line to
the tips of the waves: this shows that the cosine wave follows the x of
the end of the center line and that the sine wave follows the y of the
end of the center line.