Create gradients where all colors get brighter
As long as each component (r, g, b) are getting brighther the gradient
works. As soon as the one component is darker in the "to" color than the
"from" color the whole thing fails (returns empty array).

My goal here is to give a from color, to color and a number of
intermediate colors (i.e. length). The function should give back a list
of colors that I can use on each point in a line to create a line with a
gradient.