Add a script switcher
One of the painful things about this project was having to edit the
a.html file to try out different scripts. This adds a datalist input
field for selecting which script to run. When a script is selected the
animate function is called. Unfortunately I only started using the
my animate script later in the project so I need to retrofit the old
files (e.g. j.js) to work with the new script switcher. I _think_
whatever new script is loaded will overwrite any existing functions that
collide but I want to keep the animate() function from my animate
JavaScript script.