Move function to JavaScript specific to page
When I created this project I just kept copying the same script to a
different file and changing it. (I wanted to keep them all handy, not go
back in my commit history, or switch branches.) Because of that
everything is sloppy so I'm cleaning it up and DRYing it out.