Choose number generator function to graph
Added a drop down that grabs all the fields (which should all be
functions) from the "generators" object. The user can pick a field (i.e.
a function) and then click the "draw" button to graph a distribution of
the numbers generated by the function.

This is so I can easily experiment with functions that generate
distributions of numbers.