Add cos
cos and sin really threw me for a loop until I realized I was looking at
a graph of the _distribution_ of the numbers, not the numbers
themselves. An even distribution means a change in the input results in
in more change in the output; an uneven distribution means a change in
the input results in less of a change in the output.