Create a new animation: rotating squiggly string
I like the idea of rotating the canvas and creating a "squiggly string"
that grows out from the center. I'm using my new cos_180 function would
generates a distribution of numbers according to the cos of an angle at
increments from zero to Pi. This gives both negative and postive results
that tend towards the extremes. I'm also playing around with changing
the colour of the string. I'd like to change the individual portions of
the string but for now I'm just working on getting changing colours.
It's Christmas and I should call it quits and spend time with my family.
:)