WIP: draw non-overlapping squares
Adding a new animation that modifies the fill_with_squares to draw
squares that don't overlap each other. This entails tracking all the
squares that are drawn and all the current empty spaces that could hold
a new square.

Current I'm only drawing in the next available space, not a random
space.