Start arranging the graph for rendering
I'm going to work out all the graph vertex and edge shapes and then
render them. I'm starting from an arbitrary node (the first in the
list), putting that at the center and then working out from there.

Each arrange_shapes call gets the graph of all nodes, which is just the
list of IDs with the list of IDs each vertex is connected to, the list
of vertexes that are already arranged and accounted for, the index into
the graph for the current vertex to draw, the list of verteces and edges
already arranged, the x and y of the current vertex and the angle the
current vertex is at relative to the "parent" vertex.

I'll need to traverse the graph first to figure out the constraints:
i.e. which vertexes are connected "across" the tree instead of down:
e.g.
    1
   / \
  2---3
I need to know that 2 is connected to 3 so that I don't get:

  1---2---3 which makes the connection from 1->3 hidden.