Handle edge cases (pun intended)
The first vertex in the graph has no parent edge: when we draw the child
edges around it we don't need to leave space for the incoming parent
edge.

Edges are bi-directional so a parent edge is also a child edge: if an
angle has been specified then we know that one edge has already been
drawn and we only have to draw the rest of them. We'll know which vertex
doesn't need to be drawn because we have a list of drawn vertexes. (i.e.
every edge leads to a vertex, but for now the edges are vertex-agnostic:
we don't care which vertex the edge is leading too, we just know we need
one edge for every vertex that isn't the parent vertex, which is in the
list of drawn vertexes.)

It's possible to have a graph with one vertex and therefore zero edges.
For the "singleton" graph and for leaf nodes we don't want to calculate
the "delta angle" by dividing Pi by zero.