Wholesale re-write of arrangement logic
I need to track:
- what are the unique edges: each edge can be drawn from either
direction; I need to track each edge in a uni-directional way and I need
to have an accurate count of all edges.
- which edges have been arranged:
  It's possible to draw all the verteces but to have missed an edge
between two verteces, the edge that isn't required to reach either
vertex.
- which verteces have been arranged:
  An edge might need to be drawn to a vertex that already exists: I need
to be able to check that the vertex already exists.
- What edges have been drawn so I know how to space out the remaining
edges. This is a work in progress as it assumes the already-arranged
verteces are contiguous so all the to-be-arranged verteces are also
contiguous.

Function argument comments:
function arrange_shapes(graph, <-- map of vertex IDs to edges
                        shapes, <-- arranged shape results (for recursion)
                        uniqueEdges, <-- unique edges as 'min-max'
                        previousId, <-- ID of vertex that we're drawing edges for
                        arrangedEdgeIds, <-- edges ... this makes no sense
                        arrangedVertexIds, <-- verteces that we've already drawn
                        previousX, <-- x coord of source vertex
                        previousY, <-- y coord of source vertex
                        angle){ <-- angle from source vertex to sibling vertex