Setup for arrange loop
In order to know how to space out the yet-to-be-arranged edges we need
to know how many edges have already been arrange. This will show what
area around the current vertex is already used. The used area is stored
in "paddingAngle".

Create a new array to hold shapes draw during the current iteration.

Create a local copy of the arranged edges and verteces sets so that
we're not modifying the incoming sets. I'm being careful with the
recursion to make sure I'm not modifying anything outside of the local
scope.