Sibling edges shouldn't be ignored
When drawing sibling verteces the arranging shouldn't stop when all
verteces are arranged because there might be an edge between two
arranged verteces.

If A, B and C are all connected and we do a depth-first arrangement of
A-B, B-C, then even though all the verteces are drawn we still need the
edge between A and C.

This is broken becase now verteces are getting drawn more than once.