WIP: reduce the graph with arc accumulators
Recursively work through the graph creating vertex shapes and edges
starting from one point and working "outwards". This is redundant
because verteces only connect "outward". (Well, I think "reflexive"
would mean the verteces could connect to themselves, but this graph is
not reflexive [or whatever]). The point is, I'm thinking of the first
vertex as the center of the graph and then arranging all of its sibling
vertexes around it:

      2
      |
      1    (except evenly spaced)
     / \
    3   4--5
         \
          6

I need to:
- start at 1
- figure out how many sibling verteces 1 has (3: [2, 3, 4])
- calculate the angle between each sibling vertex
- calculate {x,y} for each vertex using that angle
- create an arc with an starting angle and arc angle for each sibling
vertex
- recurse through each siblings' siblings (e.g. 4 has [5, 6])
  - pass the "parent" arc starting angle and arc angle:
      2 would get PI * 1/6, 2PI/3 (Tau/3)
      3 would get PI * 5/6, 2PI/3
      4 would get PI * 9/6, 2PI/3
  - Now 5 and 6 will have to fit between (PI * 9/6) and (PI * 13/6)
    (270 and 30 degrees)

So, as I recurse through the graph I need to keep track of the shrinking
arc that each branch gets and what angle each vertex will use to
position itself. At this point I could also calculate the positions of
the vertexes.

I _could_ work through the entire graph first and balance it so that
more populated branches get wider arcs. In the above example vertex 4
might get half of the circle, while 2 and 3 only get 1/4 of the circle.

It's a work in progress. Everthing is a mess.