WIP: walk graph with back-tracking
My idea now is for each vertex then return to it's predecessor with any
siblings it has. The predecessor will then make sure all of the siblings
that are directly connected are next to each other. Once all of the
siblings for a vertex are drawn, then each sibling is drawn. Each
sibling will connect itself to any directly connected siblings.

Example:
Given the following graph

.   .
 \ /
  1-----2
   \   /
    \ /
     3
     |
     |
     4

Assuming we start from 3,
Calculate position of 3,
Call calculate with vertex 1 but passing in:
- a record of vertex 3
- the starting angle (zero)
Vertex 1 calculates it's position
Vertex 1 returns that it has a connection to 2
Call calculate with the sibling that was connected to 1, i.e. 2
  Pass in:
- the record of vertex 3
- the connection to 2
- the next angle (i.e. Math.PI / 3
Now 1 and 3 are both connected to 3 and each other and are beside each
other.
Now draw 4.
Once four is done call calculate on 1 _again_;
  Pass in that 3 is finished so 1 can recurse
  ... probably pass in the angle of 1 relative to 3 to continue in that
direction
...

Anyways, that's the idea.