WIP: get siblings ordered by connections
If I'm trying to arrange a set of verteces around a central vertex then
I'd like to order the surrounding verteces so that when they are
connected, they are beside each other.

This code goes through all the immediate siblings, finds the connections
to each other, then creates an ordered list where connected siblings
come first and are beside each other. Once we have that we can add all
the non-connected siblings: it doesn't matter what order we draw those.

There's at least one hole where if vertex with the minimum ID is
connected to multiple other verteces, it will only be beside the first
one. This is good enough for a first pass. I need to get this back to a
working state so I can see progress on the screen and then make small
improvements: make it work, make it right, make it fast.